+++ prog/src/table.hpp
@@ -8,8 +8,8 @@ struct Matrix
 {
  public:
 	Matrix(int n, int m);
-	int findMatrixDistDamerau(const std::wstring& st1, const std::wstring& st2);
-	int findMatrixDistLev(const std::wstring& st1, const std::wstring& st2);
+	size_t findMatrixDistDamerau(const std::wstring& st1, const std::wstring& st2);
+	size_t findMatrixDistLev(const std::wstring& st1, const std::wstring& st2);
 	size_t findRecurseDistDamerau(const std::wstring& st1, const std::wstring& st2);
 	size_t findRecurseDistMemDamerau(const std::wstring& st1, const std::wstring& st2);
